Move flip related E2E tests to `integration_flipper`
