+++ examples/lang_err_integration_tests/Cargo.toml
@@ -11,6 +11,9 @@ ink = { path = "../../crates/ink", default-features = false }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
 
+[dev-dependencies]
+ink_e2e = { path = "../../crates/e2e" }
+
 [lib]
 name = "integration_flipper"
 path = "lib.rs"
@@ -29,6 +32,7 @@ std = [
     "scale-info/std",
 ]
 ink-as-dependency = []
+e2e-tests = []
 
 [workspace]
 members = [
