+++ examples/lang_err_integration_tests/call_builder/lib.rs
@@ -12,9 +12,7 @@ mod call_builder {
 
     #[ink(storage)]
     #[derive(Default)]
-    pub struct CallBuilderTest {
-        value: bool,
-    }
+    pub struct CallBuilderTest {}
 
     impl CallBuilderTest {
         #[ink(constructor)]
@@ -56,28 +54,6 @@ mod call_builder {
                     unimplemented!("No other `LangError` variants exist at the moment.")
                 }
             }
-        }
-
-        /// Returns the current value of the `CallBuilder`'s boolean.
-        #[ink(message)]
-        pub fn get(&self) -> bool {
-            self.value
-        }
-
-        /// Flips the current value of the `CallBuilder`'s boolean.
-        #[ink(message)]
-        pub fn flip(&mut self) {
-            self.value = !self.value;
-        }
-
-        /// Flips the current value of the `CallBuilder`'s boolean.
-        ///
-        /// We should see the state being reverted here, no write should occur.
-        #[ink(message)]
-        #[allow(clippy::result_unit_err)]
-        pub fn err_flip(&mut self) -> Result<(), ()> {
-            self.flip();
-            Err(())
         }
     }
 
@@ -154,122 +130,6 @@ mod call_builder {
                 )
                 .await
                 .expect("Calling `flipper::get` failed");
-            let flipped_value = get_call_result
-                .value
-                .expect("Input is valid, call must not fail.");
-            assert!(flipped_value == initial_value);
-
-            Ok(())
-        }
-
-        #[ink_e2e::test]
-        async fn e2e_can_flip_correctly(
-            mut client: ink_e2e::Client<C, E>,
-        ) -> E2EResult<()> {
-            let constructor = call_builder::constructors::new();
-            let contract_acc_id = client
-                .instantiate(&mut ink_e2e::alice(), constructor, 0, None)
-                .await
-                .expect("Instantiate `call_builder` failed")
-                .account_id;
-
-            let get_call_result = client
-                .call(
-                    &mut ink_e2e::alice(),
-                    contract_acc_id.clone(),
-                    call_builder::messages::get(),
-                    0,
-                    None,
-                )
-                .await
-                .expect("Calling `get` failed");
-            let initial_value = get_call_result
-                .value
-                .expect("Input is valid, call must not fail.");
-
-            let flip_call_result = client
-                .call(
-                    &mut ink_e2e::alice(),
-                    contract_acc_id.clone(),
-                    call_builder::messages::flip(),
-                    0,
-                    None,
-                )
-                .await
-                .expect("Calling `flip` failed");
-            assert!(
-                flip_call_result.value.is_ok(),
-                "Messages now return a `Result`, which should be `Ok` here."
-            );
-
-            let get_call_result = client
-                .call(
-                    &mut ink_e2e::alice(),
-                    contract_acc_id.clone(),
-                    call_builder::messages::get(),
-                    0,
-                    None,
-                )
-                .await
-                .expect("Calling `get` failed");
-            let flipped_value = get_call_result
-                .value
-                .expect("Input is valid, call must not fail.");
-            assert!(flipped_value != initial_value);
-
-            Ok(())
-        }
-
-        #[ink_e2e::test]
-        async fn e2e_message_error_reverts_state(
-            mut client: ink_e2e::Client<C, E>,
-        ) -> E2EResult<()> {
-            let constructor = call_builder::constructors::new();
-            let contract_acc_id = client
-                .instantiate(&mut ink_e2e::bob(), constructor, 0, None)
-                .await
-                .expect("instantiate failed")
-                .account_id;
-
-            let get_call_result = client
-                .call(
-                    &mut ink_e2e::bob(),
-                    contract_acc_id.clone(),
-                    call_builder::messages::get(),
-                    0,
-                    None,
-                )
-                .await
-                .expect("Calling `get` failed");
-            let initial_value = get_call_result
-                .value
-                .expect("Input is valid, call must not fail.");
-
-            let err_flip_call_result = client
-                .call(
-                    &mut ink_e2e::bob(),
-                    contract_acc_id.clone(),
-                    call_builder::messages::err_flip(),
-                    0,
-                    None,
-                )
-                .await;
-
-            assert!(matches!(
-                err_flip_call_result,
-                Err(ink_e2e::Error::CallExtrinsic(_))
-            ));
-
-            let get_call_result = client
-                .call(
-                    &mut ink_e2e::bob(),
-                    contract_acc_id.clone(),
-                    call_builder::messages::get(),
-                    0,
-                    None,
-                )
-                .await
-                .expect("Calling `get` failed");
             let flipped_value = get_call_result
                 .value
                 .expect("Input is valid, call must not fail.");
